requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
google-api-python-client==1.6.7
httplib2==0.11.3
idna==2.6
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
requests==2.18.4
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
piplapis-python